Module image_rs::signature::mechanism

Signing schemes

different signatures defination and the top level interfaces.

Design

Due to the format of policy requirement in https://github.com/containers/image/blob/main/docs/containers-policy.json.5.md#policy-requirements, a signing scheme is also treated as a policy requirement. To support different kinds of signing schemes, we use a trait SignScheme to define. The trait object will be included into [crate::policy::PolicyReqType].

Modules

cosign

Cosign verification

simple

Traits

SignScheme

The interface of a signing scheme